# 【PFCC算子性能优化】为Paddle优化adaptive_pooling_op性能