<!-- Demo: https://github.com/PaddlePaddle/Paddle/pull/24810 -->
### PR types
<!-- One of [ New features | Bug fixes | Function optimization | Performance optimization | Breaking changes | Others ] -->Performance optimization

### PR changes
<!-- One of [ OPs | APIs | Docs | Others ] -->OPs

### Describe
<!-- Describe what this PR does -->
- document: [PFCC PR 244](https://github.com/PaddlePaddle/community/pull/244)
- device: GeForce RTX 3080 (Compute Capability: 8.6)
- data_type: float32
- data_format: NCHW
- diff: (before - after) / before

| Case No. | input\_shape        | output\_shape | pooling\_type | before kernel\_time(ms) | after kernel\_time(ms) | diff    |
| -------- | ------------------- | ------------- | ------------- | ----------------------- | ---------------------- | ------- |
| 0        | \[128,64,112,112\]  | \[56,56\]     | AVG           | 5.7967                  | 0.7395                 | 87.24%  |
| 1        | \[128,512,7,7\]     | \[1,1\]       | AVG           | 0.0341                  | 0.0341                 | 0.00%   |
| 2        | \[128,2048,7,7\]    | \[1,1\]       | AVG           | 0.1239                  | 0.1242                 | \-0.24% |
| 3        | \[4,2048,64,128\]   | \[32,32\]     | AVG           | 2.1829                  | 0.4311                 | 80.25%  |
| 4        | \[128,64,224,224\]  | \[112,112\]   | AVG           | 19.1376                 | 3.2681                 | 82.92%  |
| 5        | \[128,128,112,112\] | \[56,56\]     | AVG           | 11.0257                 | 1.4704                 | 86.66%  |
| 6        | \[128,256,56,56\]   | \[28,28\]     | AVG           | 6.7391                  | 0.7834                 | 88.38%  |
| 7        | \[128,512,28,28\]   | \[14,14\]     | AVG           | 3.5089                  | 0.4116                 | 88.27%  |
| 8        | \[128,512,14,14\]   | \[7,7\]       | AVG           | 0.9216                  | 0.1138                 | 87.65%  |
| 9        | \[128,64,112,112\]  | \[56,56\]     | MAX           | 5.9602                  | 0.9065                 | 84.79%  |
| 10       | \[128,512,7,7\]     | \[1,1\]       | MAX           | 0.0895                  | 0.0739                 | 17.43%  |
| 11       | \[128,2048,7,7\]    | \[1,1\]       | MAX           | 0.3027                  | 0.2711                 | 10.44%  |
| 12       | \[4,2048,64,128\]   | \[32,32\]     | MAX           | 2.2829                  | 0.4811                 | 78.93%  |
| 13       | \[128,64,224,224\]  | \[112,112\]   | MAX           | 21.6426                 | 4.2093                 | 80.55%  |
| 14       | \[128,128,112,112\] | \[56,56\]     | MAX           | 11.8736                 | 1.8121                 | 84.74%  |
| 15       | \[128,256,56,56\]   | \[28,28\]     | MAX           | 6.8098                  | 0.891                  | 86.92%  |
| 16       | \[128,512,28,28\]   | \[14,14\]     | MAX           | 3.5734                  | 0.4537                 | 87.30%  |
| 17       | \[128,512,14,14\]   | \[7,7\]       | MAX           | 0.9262                  | 0.1214                 | 86.89%  |

### TODO
- [x] adaptive pooling forward optimization
- [ ] adaptive pooling backward optimization